pyproject.toml
[project]
name = "goth-aesthetics-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "mastodon-py>=2.1.4",
    "python-dotenv>=1.2.2",
]
